+++ Mathematics/nCr3.cpp
@@ -1,5 +1,5 @@
 //nCr(modを使わない)
-ll com[200][200];
+long long com[200][200];
 
 void init() {
     for (size_t i = 0; i < 200; i++) {
@@ -9,7 +9,7 @@ void init() {
     }
 }
 
-ll nCr(int n,int r) {
+long long nCr(int n,int r) {
     if(0 > n || r > n) return 0;
     if(n == 0 || r == 0) return com[n][r] = 1;
     if(com[n][r] != -1) return com[n][r];
